Generacion de Permisos de Depositos

+++ app/Http/Controllers/admin/stock/StoreController.php
@@ -43,6 +43,7 @@ class StoreController extends Controller
 
     public function create()
     {   
+        $this->authorize('create', 'App\Models\Store');
         $storeTypeList = StoreType::all();
         return view('dashboard.stock.stores.create', compact('storeTypeList'));
     }
@@ -72,6 +73,7 @@ class StoreController extends Controller
 
     public function edit($id){
         $store = $this->storeRepository->find($id);
+        $this->authorize('update', $store);
         $storeTypeList = StoreType::all();
         return view('dashboard.stock.stores.edit', compact('store', 'storeTypeList'));
     }
@@ -99,6 +101,8 @@ class StoreController extends Controller
 
 
     public function destroy($id){
+        $store = $this->storeRepository->find($id);
+        $this->authorize('delete', $store);
         $this->storeRepository->delete($id);
         return response()
         ->json([
